streamlit_demo/requirements.txt
streamlit>=1.31
openai>=1.40.0
python-dotenv>=1.0.1
requests>=2.31.0
schedule>=1.2.0
numpy
numpy
